Move validator creation helpers outside of test
