+++ packages/shared/src/nomination-pools.ts
@@ -2,14 +2,22 @@ import { encodeAddress } from '@polkadot/util-crypto'
 
 import { type Chain, defaultAccounts } from '@e2e-test/networks'
 import { setupNetworks } from '@e2e-test/shared'
-import { check, checkEvents, checkSystemEvents, objectCmp, scheduleCallWithOrigin } from './helpers/index.js'
+import {
+  check,
+  checkEvents,
+  checkSystemEvents,
+  createAndBondAccounts,
+  getValidators,
+  objectCmp,
+  scheduleCallWithOrigin,
+  setValidatorsStorage,
+} from './helpers/index.js'
 
-import type { StorageValues } from '@acala-network/chopsticks'
 import { sendTransaction } from '@acala-network/chopsticks-testing'
 import type { ApiPromise } from '@polkadot/api'
 import type { KeyringPair } from '@polkadot/keyring/types'
 import { type Option, u32 } from '@polkadot/types'
-import type { PalletNominationPoolsBondedPoolInner, PalletStakingValidatorPrefs } from '@polkadot/types/lookup'
+import type { PalletNominationPoolsBondedPoolInner } from '@polkadot/types/lookup'
 import type { Codec } from '@polkadot/types/types'
 import { assert, describe, test } from 'vitest'
 
@@ -74,131 +82,6 @@ async function createNominationPool(
   const createNomPoolEvents = sendTransaction(createNomPoolTx.signAsync(signer))
 
   return createNomPoolEvents
-}
-
-/**
- * Select some validators from the list present in the `Validators` storage item, in the `Staking` pallet.
- *
- * To avoid fetching all validators at once (over a thousand in Jan. 2025), only the first page of validators
- * in storage is considered - the size of the page is provided as an argument.
- *
- * If, in the validator page of the selected size, less than `validatorCount` validators are available, the function
- * will get as close to `validatorCount` as possible.
- *
- * @param api PJS client object.
- * @param pageSize The size of the page to fetch from storage.
- * @param validatorCount The (desired) number of validators to select.
- * @returns A list of at least 1 validator, and at most 16.
- */
-async function getValidators(api: ApiPromise, pageSize: number, validatorCount: number): Promise<string[]> {
-  // Between 1 and 16 validators can be nominated by the pool at any time.
-  const min_validators = 1
-  const max_validators = 16
-
-  assert(pageSize >= max_validators)
-  assert(min_validators <= validatorCount && validatorCount <= max_validators)
-
-  // Query the list of validators from the `Validators` storage item in the `staking` pallet.
-  const validators = await api.query.staking.validators.entriesPaged({ args: [], pageSize: pageSize })
-
-  const validatorIds: [string, PalletStakingValidatorPrefs][] = validators.map((tuple) => [
-    tuple[0].args[0].toString(),
-    tuple[1],
-  ])
-
-  const selectedValidators: string[] = []
-
-  let ix = 0
-  let count = 0
-  while (count < validatorCount) {
-    const [valAddr, valData] = validatorIds[ix]
-
-    // The pool's nominator should only select validators who still allow for nominators
-    // to select them i.e. they have not blocked themselves.
-    if (valData.blocked.isFalse) {
-      selectedValidators.push(valAddr)
-      count += 1
-    }
-
-    ix += 1
-  }
-
-  assert(selectedValidators.length >= min_validators && selectedValidators.length <= max_validators)
-
-  return selectedValidators
-}
-
-/**
- * Create a given number of keypairs, add some funds to them, and bond those funds.
- */
-async function createAndBondAccounts(
-  client: {
-    api: ApiPromise
-    dev: {
-      newBlock: (param?: any) => Promise<string>
-      setStorage: (values: StorageValues, blockHash?: string) => Promise<any>
-    }
-  },
-  validatorCount: number,
-): Promise<KeyringPair[]> {
-  const validators: KeyringPair[] = []
-
-  for (let i = 0; i < validatorCount; i++) {
-    const validator = defaultAccounts.keyring.addFromUri(`//Validator_${i}`)
-    validators.push(validator)
-  }
-
-  await client.dev.setStorage({
-    System: {
-      account: validators.map((v) => [[v.address], { providers: 1, data: { free: 10000e10 } }]),
-    },
-  })
-
-  for (let i = 0; i < validatorCount; i++) {
-    const bondTx = client.api.tx.staking.bond(1000e10, { Staked: null })
-    await sendTransaction(bondTx.signAsync(validators[i]))
-  }
-
-  await client.dev.newBlock()
-
-  return validators
-}
-
-/**
- * Insert the given validators into storage.
- *
- * The `Validators` storage item is *not* meant to be manipulated directly.
- * However, in the case that the test chain has no validators and it is impracticable to call `validate` and wait
- * for the next era, this function can be used.
- *
- * Note also that normally, a successful call to `validate` would also manipulate the `VoterList` in storage, which is
- * not done here
- * For the purposes of most tests (e.g. just verifying that nominating existing validators works), this can be ignored.
- * @param client
- * @param validators
- */
-async function setValidatorsStorage(
-  client: {
-    api: ApiPromise
-    dev: {
-      setStorage: (values: StorageValues, blockHash?: string) => Promise<any>
-    }
-  },
-  validators: string[],
-) {
-  const minCommission = await client.api.query.staking.minCommission()
-
-  await client.dev.setStorage({
-    Staking: {
-      Validators: validators.map((val) => [
-        [val],
-        {
-          blocked: false,
-          commission: minCommission,
-        },
-      ]),
-    },
-  })
 }
 
 /// -------
